pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "kdx"
version = "0.1.0"
description = "AI-powered Kubernetes deployment failure diagnosis"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "pydantic>=2.5",
    "anthropic>=0.25",
    "kubernetes>=29.0",
    "python-dotenv>=1.0",
]

[project.scripts]
kdx = "kdx.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["kdx*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["kdx"]
omit = ["tests/*"]
